Estimate the unknown parameters of the admixture model(s)

Estimate the component weights, the location shift parameter (in case of a symmetric unknown component density), and the unknown component distribution using different estimation techniques. We remind that the i-th admixture model has probability density function (pdf) l_i such that: l_i = p_i * f_i + (1-p_i) * g_i, where g_i is the known component density. The unknown quantities p_i and f_i then have to be estimated.

Usage

admix_estim(
  samples,
  admixMod,
  est_method = c("BVdk", "PS", "IBM"),
  sym_f = FALSE
)

Arguments

samples

(list) A list of the K (K>0) samples to be studied, all following admixture distributions.

admixMod

(list) A list of objects of class 'admix_model', containing useful information about distributions and parameters.

est_method

The estimation method to be applied. Can be one of 'BVdk' (Bordes and Vandekerkhove estimator), 'PS' (Patra and Sen estimator), or 'IBM' (Inversion Best-Matching approach). The same estimation method is performed on each sample. Important note: estimation by 'IBM' is unbiased only under H0, meaning that choosing this method requires to perform previously the test hypothesis between the pairs of samples. For further details, see section 'Details' below.

sym_f

A boolean indicating whether the unknown component densities are assumed to be symmetric or not.

Value

An object of class 'admix_estim', containing at least 5 attributes: 1) the number of samples under study; 2) the information about the mixture components (distributions and parameters); 3) the sizes of the samples; 4) the chosen estimation technique (one of 'BVdk', 'PS' or 'IBM'); 5) the estimated mixing proportions (weights of the unknown component distributions in the mixture model). In case of 'BVdk' estimation, one additional attribute corresponding to the estimated location shift parameter is included.

Details

For further details on the different estimation techniques, see references below i) Patra and Sen estimator ; ii) BVdk estimator ; iii) IBM approach.

References

PatraSen2016admix BordesDelmasVandekerkhove2006admix BordesVandekerkhove2010admix MilhaudPommeretSalhiVandekerkhove2024aadmix

Author

Xavier Milhaud xavier.milhaud.research@gmail.com

Examples

## Simulate mixture data:
mixt1 <- twoComp_mixt(n = 250, weight = 0.7,
                      comp.dist = list("norm", "norm"),
                      comp.param = list(list("mean" = -2, "sd" = 0.5),
                                        list("mean" = 0, "sd" = 1)))
mixt2 <- twoComp_mixt(n = 200, weight = 0.85,
                      comp.dist = list("norm", "exp"),
                      comp.param = list(list("mean" = 3, "sd" = 1),
                                        list("rate" = 1)))
data1 <- getmixtData(mixt1)
data2 <- getmixtData(mixt2)

## Define the admixture models:
admixMod1 <- admix_model(knownComp_dist = mixt1$comp.dist[[2]],
                         knownComp_param = mixt1$comp.param[[2]])
admixMod2 <- admix_model(knownComp_dist = mixt2$comp.dist[[2]],
                         knownComp_param = mixt2$comp.param[[2]])
admix_estim(samples = list(data1), admixMod = list(admixMod1),
            est_method = 'BVdk', sym_f = TRUE)
#> Call:
#> admix_estim(samples = list(data1), admixMod = list(admixMod1), 
#>     est_method = "BVdk", sym_f = TRUE)
#> 
#> Estimated mixing weight of the unknown component distribution in Sample 1: 0.69
#> 
admix_estim(samples = list(data1,data2),
            admixMod = list(admixMod1,admixMod2),
            est_method = 'PS')
#> Call:
#> admix_estim(samples = list(data1, data2), admixMod = list(admixMod1, 
#>     admixMod2), est_method = "PS")
#> 
#> Estimated mixing weight of the unknown component distribution in Sample 1: 0.65
#> Estimated mixing weight of the unknown component distribution in Sample 2: 0.76
#>