Estimates weights of unknown components from 2 admixtures using IBM

Estimation of the component weights from the Inversion - Best Matching (IBM) method, related to two admixture models with respective probability density function (pdf) l1 and l2, such that: l1 = p1*f1 + (1-p1)g1 and l2 = p2f2 + (1-p2)*g2, where g1 and g2 are the known component densities. For further details about IBM approach, see 'Details' below.

Usage

estim_IBM(samples, admixMod, n.integ = 1000, compute_var = FALSE)

Arguments

samples

A list of the two considered samples.

admixMod

A list of two objects of class admix_model, one for each sample.

n.integ

Number of data points generated for the distribution on which to integrate.

compute_var

(default to FALSE) A boolean that indicates whether one computes the variance of the estimators of unknown mixing proportions.

Value

An object of class estim_IBM, containing 7 attributes: 1) the number of samples under study; 2) the sizes of samples; 3) the information about mixture components (distributions and parameters) for each sample; 4) the estimation method (Inversion Best Matching here, see the given reference); 5) the estimated mixing proportions (weights of the unknown component distributions in each sample); 6) the arbitrary value of the mixing weight in the first admixture sample (in case of equal known components, see the given reference); 7) the support of integration that was used in the computations.

References

Milhaud X, Pommeret D, Salhi Y, Vandekerkhove P (2024). “Two-sample contamination model test.” Bernoulli, 30(1), 170–197. doi:10.3150/23-BEJ1593 .

See also

Author

Xavier Milhaud xavier.milhaud.research@gmail.com

Examples

## Continuous support: simulate mixture data.
mixt1 <- twoComp_mixt(n = 1500, weight = 0.5,
                      comp.dist = list("norm", "norm"),
                      comp.param = list(list("mean" = 3, "sd" = 0.5),
                                        list("mean" = 0, "sd" = 1)))
data1 <- getmixtData(mixt1)
mixt2 <- twoComp_mixt(n = 2000, weight = 0.7,
                      comp.dist = list("norm", "norm"),
                      comp.param = list(list("mean" = 3, "sd" = 0.5),
                                        list("mean" = 5, "sd" = 2)))
data2 <- getmixtData(mixt2)
## Define the admixture models:
admixMod1 <- admix_model(knownComp_dist = mixt1$comp.dist[[2]],
                         knownComp_param = mixt1$comp.param[[2]])
admixMod2 <- admix_model(knownComp_dist = mixt2$comp.dist[[2]],
                         knownComp_param = mixt2$comp.param[[2]])
## Estimate the mixture weights of the two admixture models (provide only hat(theta)_n):
estim_IBM(samples = list(data1,data2),
          admixMod = list(admixMod1,admixMod2), n.integ = 1000)
#> Warning:  IBM estimators of two unknown proportions are reliable only if the two
#>     corresponding unknown component distributions have been tested equal (see 'admix_test()').
#>     Furthermore, when both the known and unknown component distributions of the mixture
#>     models are identical, the IBM approach provides an estimation of the ratio of the
#>     actual mixing weights rather than an estimation of the unknown weights themselves.
#> $n_populations
#> [1] 2
#> 
#> $population_sizes
#> [1] 1500 2000
#> 
#> $admixture_models
#> $admixture_models[[1]]
#> Call:admix_model(knownComp_dist = mixt1$comp.dist[[2]], knownComp_param = mixt1$comp.param[[2]])
#> 
#> Known component distribution:  norm 
#> Value of the known component parameters:
#> $mean
#> [1] 0
#> 
#> $sd
#> [1] 1
#> 
#> 
#> 
#> $admixture_models[[2]]
#> Call:admix_model(knownComp_dist = mixt2$comp.dist[[2]], knownComp_param = mixt2$comp.param[[2]])
#> 
#> Known component distribution:  norm 
#> Value of the known component parameters:
#> $mean
#> [1] 5
#> 
#> $sd
#> [1] 2
#> 
#> 
#> 
#> 
#> $estimation_method
#> [1] "Inversion Best Matching (IBM)"
#> 
#> $estimated_mixing_weights
#> [1] 0.5111508 0.7109760
#> 
#> $variance_est_p1
#> [1] NA
#> 
#> $variance_est_p2
#> [1] NA
#> 
#> $equal.knownComp
#> [1] FALSE
#> 
#> $p.X.fixed
#> NULL
#> 
#> $integ.supp
#>    [1] -3.177482267 -3.150650472 -3.140770947 -3.135009339 -3.126353868
#>    [6] -3.125158426 -3.112818378 -3.079682017 -3.077588998 -3.068663954
#>   [11] -3.057174978 -3.055763960 -3.018871409 -3.010612124 -2.987284014
#>   [16] -2.982621240 -2.982387061 -2.969573341 -2.965173645 -2.937420094
#>   [21] -2.912342049 -2.903329645 -2.896371431 -2.891853957 -2.888146693
#>   [26] -2.886987261 -2.885280263 -2.884339763 -2.871546408 -2.859345733
#>   [31] -2.852457980 -2.849814852 -2.836899292 -2.819773722 -2.815184040
#>   [36] -2.813441518 -2.810655024 -2.803395683 -2.788372166 -2.785662001
#>   [41] -2.783446616 -2.774390223 -2.772327844 -2.760844460 -2.755929967
#>   [46] -2.753892327 -2.748121388 -2.714872214 -2.708523612 -2.642117563
#>   [51] -2.614577596 -2.611421961 -2.608897032 -2.589479640 -2.588644368
#>   [56] -2.586267262 -2.583818919 -2.572893175 -2.572308697 -2.553008093
#>   [61] -2.542699772 -2.530785879 -2.527016840 -2.513543422 -2.495220348
#>   [66] -2.483543692 -2.463578118 -2.463386247 -2.460077705 -2.454572062
#>   [71] -2.432878763 -2.377633442 -2.365777174 -2.363758260 -2.348833397
#>   [76] -2.348775614 -2.327789703 -2.326022327 -2.320570371 -2.319993424
#>   [81] -2.316112856 -2.313343984 -2.305707341 -2.288271182 -2.276829257
#>   [86] -2.271762847 -2.271551964 -2.270797588 -2.228833908 -2.196461039
#>   [91] -2.163426659 -2.157212746 -2.149019542 -2.144504619 -2.112413226
#>   [96] -2.097762235 -2.089610641 -2.084655754 -2.039081556 -2.017453634
#>  [101] -2.017220189 -1.997209090 -1.992011529 -1.989934623 -1.957239268
#>  [106] -1.942002839 -1.936034471 -1.911894809 -1.908572990 -1.907655091
#>  [111] -1.904712664 -1.904085181 -1.894974761 -1.894890895 -1.884538983
#>  [116] -1.875139118 -1.864398542 -1.823115247 -1.814639304 -1.785420935
#>  [121] -1.775690665 -1.765785078 -1.753793241 -1.746444173 -1.740574015
#>  [126] -1.740480485 -1.720759700 -1.667029640 -1.666130224 -1.654149721
#>  [131] -1.633945015 -1.631649257 -1.591330158 -1.549679926 -1.541821869
#>  [136] -1.527122129 -1.512215552 -1.511855083 -1.503805432 -1.501618845
#>  [141] -1.453802763 -1.448117319 -1.432320128 -1.423735302 -1.409822344
#>  [146] -1.393646309 -1.359606677 -1.340472251 -1.328274503 -1.322613335
#>  [151] -1.319775130 -1.280536799 -1.278510276 -1.233782384 -1.215743714
#>  [156] -1.209876749 -1.207206403 -1.166371107 -1.163247502 -1.148256894
#>  [161] -1.135346004 -1.135312605 -1.132176496 -1.118175843 -1.089237957
#>  [166] -1.064116555 -1.060117241 -1.057959172 -1.056107765 -1.036986889
#>  [171] -1.030542097 -1.023371302 -1.014880200 -1.014417579 -1.003325805
#>  [176] -0.990952033 -0.971784199 -0.963511439 -0.959550517 -0.936354318
#>  [181] -0.919057791 -0.916929421 -0.916448125 -0.911677811 -0.911201339
#>  [186] -0.909247947 -0.908832729 -0.861478848 -0.853858479 -0.823476790
#>  [191] -0.805982460 -0.804788831 -0.796160355 -0.780738830 -0.774423068
#>  [196] -0.769777586 -0.765439356 -0.762411274 -0.755039101 -0.736255060
#>  [201] -0.695061594 -0.694038533 -0.661981466 -0.644311759 -0.635804867
#>  [206] -0.611781352 -0.607182596 -0.568600892 -0.549651722 -0.518938278
#>  [211] -0.508884165 -0.508558154 -0.494671598 -0.487377500 -0.459695745
#>  [216] -0.457116458 -0.411460778 -0.400091205 -0.382082914 -0.352318869
#>  [221] -0.350185250 -0.348133424 -0.304780212 -0.268040012 -0.222203062
#>  [226] -0.215029748 -0.161439978 -0.149111116 -0.143803568 -0.141325848
#>  [231] -0.140268014 -0.113667208 -0.103895689 -0.074791737 -0.057878435
#>  [236] -0.047839264 -0.036006165 -0.030039902 -0.029856057 -0.028746531
#>  [241] -0.023656500 -0.020986039 -0.002406057  0.069227573  0.079859938
#>  [246]  0.093272226  0.093335889  0.100629698  0.101621136  0.114220694
#>  [251]  0.124785676  0.167860465  0.193121963  0.211989960  0.212968591
#>  [256]  0.216070131  0.220449732  0.288301746  0.294545864  0.298093031
#>  [261]  0.300511442  0.303165098  0.314502077  0.349842540  0.357634826
#>  [266]  0.370958149  0.395324501  0.404694259  0.407670347  0.408077344
#>  [271]  0.424982749  0.444842628  0.475016827  0.506982506  0.522612994
#>  [276]  0.574215479  0.575474143  0.585022550  0.594921806  0.612715737
#>  [281]  0.623864922  0.628030893  0.628927154  0.631983445  0.632682459
#>  [286]  0.638717913  0.645650316  0.657256922  0.670844738  0.699882484
#>  [291]  0.701987593  0.712724423  0.717022986  0.743149843  0.749616560
#>  [296]  0.759664591  0.767118339  0.772380760  0.798295047  0.800660329
#>  [301]  0.801485764  0.813871587  0.814526233  0.816147589  0.832144064
#>  [306]  0.836768421  0.844458544  0.848397000  0.866376366  0.882042449
#>  [311]  0.892617528  0.907196832  0.907320274  0.916274543  0.922828800
#>  [316]  0.934331921  0.954493739  0.980699566  0.985218192  0.986224960
#>  [321]  1.003831356  1.016764817  1.030338284  1.044227746  1.075792059
#>  [326]  1.111647657  1.137972074  1.147063112  1.155928717  1.170677496
#>  [331]  1.225563848  1.226987337  1.233391140  1.239600606  1.248494541
#>  [336]  1.253196397  1.258323208  1.339010014  1.353871928  1.369221446
#>  [341]  1.395170039  1.410936700  1.417040409  1.426043814  1.427272724
#>  [346]  1.427773572  1.434335551  1.442993258  1.456908594  1.460443454
#>  [351]  1.513262190  1.542263667  1.550448238  1.550807849  1.557910826
#>  [356]  1.573678698  1.583039651  1.587373379  1.589437258  1.604729671
#>  [361]  1.611027223  1.622982596  1.644832038  1.675751788  1.695186325
#>  [366]  1.709130762  1.746749525  1.778246350  1.780475361  1.814700348
#>  [371]  1.822082569  1.827754328  1.829866777  1.833873402  1.839958505
#>  [376]  1.852477185  1.872273256  1.893800463  1.913118721  1.925856317
#>  [381]  1.959385447  1.970755020  1.975762329  2.002314150  2.042441075
#>  [386]  2.056510828  2.062459653  2.062544164  2.065432488  2.067044115
#>  [391]  2.072325615  2.075592645  2.101771088  2.114391978  2.157135779
#>  [396]  2.166721798  2.171503373  2.176422411  2.183203032  2.191224941
#>  [401]  2.198068172  2.262194788  2.274548988  2.287475928  2.300366154
#>  [406]  2.304097374  2.309082690  2.316463276  2.318915780  2.320354527
#>  [411]  2.343659051  2.364236303  2.382771253  2.426913272  2.452534059
#>  [416]  2.454989302  2.455974094  2.457944470  2.471087805  2.496564423
#>  [421]  2.500621315  2.504288984  2.518344818  2.551513674  2.553151708
#>  [426]  2.604829582  2.610473168  2.617142782  2.627011988  2.643445579
#>  [431]  2.657092766  2.668626470  2.690099022  2.706025668  2.722411856
#>  [436]  2.729920268  2.729997567  2.733778814  2.762261716  2.805056416
#>  [441]  2.819194768  2.822153616  2.865937182  2.892010178  2.894258892
#>  [446]  2.931431518  2.932167328  2.954952978  2.960686447  2.992732509
#>  [451]  3.019626829  3.044769010  3.056933855  3.068416191  3.078276072
#>  [456]  3.088760589  3.108536815  3.124198748  3.126636251  3.127701031
#>  [461]  3.127736716  3.147910279  3.148846892  3.159309350  3.167813739
#>  [466]  3.188297224  3.221515268  3.222786828  3.231478180  3.235661830
#>  [471]  3.241728111  3.256097889  3.257349922  3.266790283  3.282796697
#>  [476]  3.286501201  3.304338937  3.307740298  3.341919579  3.354602639
#>  [481]  3.357376036  3.363393355  3.379398529  3.385501583  3.402184242
#>  [486]  3.421950026  3.422969821  3.424865510  3.433471598  3.440075572
#>  [491]  3.465100241  3.479472095  3.495727882  3.516840377  3.553507008
#>  [496]  3.580267729  3.581355601  3.590863433  3.595867203  3.620047072
#>  [501]  3.663716527  3.685614908  3.706391985  3.727268739  3.736427603
#>  [506]  3.754385370  3.772275215  3.779074093  3.787356975  3.795009608
#>  [511]  3.801340924  3.820022207  3.843032196  3.845424283  3.879539128
#>  [516]  3.905098057  3.913461971  3.929250734  3.930934904  3.940775022
#>  [521]  3.951906295  3.968711087  3.984569544  4.012395418  4.020054464
#>  [526]  4.025197660  4.047721116  4.048204425  4.056133834  4.072622998
#>  [531]  4.081753342  4.087066493  4.103086925  4.131316562  4.146960419
#>  [536]  4.177309659  4.235131455  4.277262901  4.287176773  4.296365458
#>  [541]  4.301429210  4.321195247  4.326688166  4.329187507  4.329456785
#>  [546]  4.330390414  4.347286897  4.355830523  4.368550917  4.368932994
#>  [551]  4.378518917  4.378911896  4.407551842  4.423275781  4.444373965
#>  [556]  4.446015008  4.457849488  4.470220863  4.471798578  4.482128224
#>  [561]  4.489963588  4.502765290  4.504512292  4.521732572  4.533107048
#>  [566]  4.548770622  4.587934907  4.608550003  4.613644960  4.616058665
#>  [571]  4.619003028  4.638566246  4.652789695  4.668444735  4.672640477
#>  [576]  4.677666243  4.700315779  4.721270973  4.735425039  4.745920521
#>  [581]  4.748636988  4.750264701  4.753272010  4.770437863  4.772130548
#>  [586]  4.772696704  4.839540888  4.855729865  4.874941761  4.882478914
#>  [591]  4.893079956  4.898798582  4.903962550  4.906383859  4.909471907
#>  [596]  4.912846128  4.924752845  4.928575230  4.928855104  4.956018599
#>  [601]  4.967875201  4.998353730  5.023398981  5.032795116  5.048515449
#>  [606]  5.064623247  5.072311482  5.078384709  5.089488119  5.095260429
#>  [611]  5.118828170  5.122813125  5.124860944  5.145968985  5.152500742
#>  [616]  5.154992693  5.193459306  5.194089608  5.205689412  5.256663441
#>  [621]  5.330827393  5.373358469  5.374304413  5.398657182  5.407595342
#>  [626]  5.414695327  5.430713190  5.465307394  5.473298460  5.477583571
#>  [631]  5.485021996  5.495237728  5.529650474  5.532760394  5.560172601
#>  [636]  5.612436982  5.624422830  5.661666690  5.662494536  5.670609501
#>  [641]  5.673542173  5.674194925  5.678259687  5.704158789  5.711766358
#>  [646]  5.716141259  5.723782815  5.725529619  5.760540341  5.769859592
#>  [651]  5.781480843  5.783568721  5.826635143  5.839073143  5.854314565
#>  [656]  5.859446105  5.884886723  5.893945665  5.944191784  5.963209194
#>  [661]  6.016798188  6.021846999  6.038049891  6.040613306  6.058050403
#>  [666]  6.083597502  6.120862397  6.123334907  6.128804489  6.134579364
#>  [671]  6.158118275  6.164098927  6.179272746  6.183320048  6.189175346
#>  [676]  6.211283019  6.249130064  6.259574848  6.271027234  6.282531905
#>  [681]  6.293428157  6.312484015  6.326156779  6.343828788  6.363762568
#>  [686]  6.376776322  6.393058105  6.393339771  6.411756773  6.418393350
#>  [691]  6.422197136  6.429007574  6.438798655  6.455874891  6.528246840
#>  [696]  6.531080368  6.532071929  6.532348725  6.533178699  6.544852010
#>  [701]  6.551137909  6.556116266  6.557396548  6.583085069  6.596832994
#>  [706]  6.620199153  6.658130153  6.685165310  6.699416543  6.701676733
#>  [711]  6.716568842  6.739373920  6.744898639  6.815520489  6.824415720
#>  [716]  6.827514845  6.834380408  6.837081335  6.845922999  6.850773273
#>  [721]  6.854063901  6.854626952  6.857243295  6.861735353  6.886021752
#>  [726]  6.921040278  6.944884250  6.952350077  6.953959337  6.961078121
#>  [731]  6.964851541  6.974325116  6.977098575  6.980800560  6.989971281
#>  [736]  7.011284816  7.034316113  7.049344244  7.061377215  7.072122350
#>  [741]  7.089042516  7.121318297  7.127101267  7.140338944  7.145527760
#>  [746]  7.149173459  7.178173288  7.188622949  7.291601885  7.292473591
#>  [751]  7.303256930  7.305874819  7.313284791  7.327518429  7.334524398
#>  [756]  7.335807222  7.348951692  7.357446632  7.366963588  7.367215809
#>  [761]  7.392919827  7.398609767  7.424388569  7.429499516  7.431380622
#>  [766]  7.434286880  7.442307316  7.449662358  7.452453449  7.464793688
#>  [771]  7.513348976  7.513936069  7.530067280  7.531477055  7.542585494
#>  [776]  7.542840336  7.557495504  7.566082363  7.591757169  7.616329994
#>  [781]  7.633883011  7.637505093  7.697858289  7.718498555  7.723286827
#>  [786]  7.726460712  7.735617337  7.738546055  7.753396251  7.763088642
#>  [791]  7.779531574  7.785787841  7.785859552  7.786428197  7.794208133
#>  [796]  7.806911583  7.827633739  7.880721937  7.889888281  7.898460655
#>  [801]  7.903124182  7.910045804  7.965861690  7.969222466  7.974245062
#>  [806]  7.977790940  7.986860159  7.986895403  7.993761551  8.007061714
#>  [811]  8.014249598  8.021652745  8.030134797  8.034091621  8.034228863
#>  [816]  8.043428040  8.069628404  8.115355765  8.119341506  8.120333415
#>  [821]  8.130959547  8.154767078  8.174140600  8.196846665  8.199242916
#>  [826]  8.206682469  8.264519498  8.273176705  8.284417860  8.284741398
#>  [831]  8.286509339  8.286872590  8.357881680  8.423933358  8.437453810
#>  [836]  8.454985210  8.455175682  8.465854105  8.486072239  8.488032867
#>  [841]  8.525358002  8.537231889  8.546656174  8.560128148  8.581550265
#>  [846]  8.589705621  8.590223864  8.647945944  8.648975619  8.658475462
#>  [851]  8.681188908  8.706698945  8.707015671  8.759258437  8.761750053
#>  [856]  8.776043364  8.800750656  8.813516115  8.825124701  8.834632385
#>  [861]  8.840785657  8.842688660  8.849823309  8.853346047  8.861364154
#>  [866]  8.875329393  8.944881436  9.023553191  9.037991365  9.068538487
#>  [871]  9.088326356  9.091419078  9.102560520  9.116278214  9.125701959
#>  [876]  9.146418455  9.168243766  9.180026019  9.189618037  9.207287316
#>  [881]  9.210094118  9.214420952  9.223831200  9.231965826  9.246351667
#>  [886]  9.256743779  9.292778082  9.317153409  9.346297357  9.375746574
#>  [891]  9.381840435  9.384587367  9.446890521  9.502326449  9.515905283
#>  [896]  9.538112329  9.544544745  9.544677422  9.580003650  9.595354444
#>  [901]  9.597862534  9.628206528  9.644489860  9.693170932  9.756515868
#>  [906]  9.768654593  9.778604343  9.786192544  9.787943396  9.788655973
#>  [911]  9.792051217  9.809806148  9.829305272  9.844681335  9.877554359
#>  [916]  9.911112537  9.921296785  9.928112429  9.949335374  9.949902145
#>  [921]  9.958897459  9.969006961  9.969958518 10.001845093 10.011998408
#>  [926] 10.018579933 10.029496813 10.041534691 10.054168943 10.064549281
#>  [931] 10.158937462 10.159622708 10.178142872 10.181482997 10.203817780
#>  [936] 10.208836399 10.218000253 10.232285497 10.232429642 10.253914001
#>  [941] 10.258482881 10.274134119 10.280524947 10.285512216 10.286451943
#>  [946] 10.293504341 10.346674771 10.350877427 10.365978841 10.375828435
#>  [951] 10.377929136 10.383517585 10.431532996 10.432267075 10.432831577
#>  [956] 10.434972764 10.440128494 10.458140544 10.466568710 10.470596699
#>  [961] 10.473689740 10.491209664 10.492534977 10.501278265 10.507914773
#>  [966] 10.508532512 10.510606343 10.511502831 10.574885928 10.607358984
#>  [971] 10.654863443 10.670588460 10.670692918 10.676246639 10.678155970
#>  [976] 10.689933204 10.701046151 10.710307557 10.714471723 10.737918962
#>  [981] 10.744684401 10.749261411 10.754072821 10.757684756 10.760883314
#>  [986] 10.764260532 10.777255203 10.783535939 10.787954786 10.822444890
#>  [991] 10.823640033 10.831572205 10.837566025 10.838662777 10.855932229
#>  [996] 10.862106833 10.897859824 10.902067189 10.905326875 10.926080542
#> 
#> $call
#> estim_IBM(samples = list(data1, data2), admixMod = list(admixMod1, 
#>     admixMod2), n.integ = 1000)
#> 
#> attr(,"class")
#> [1] "estim_IBM"   "admix_estim"

## Example 2: multinomial distribution (discrete case).
mixt1 <- twoComp_mixt(n = 1500, weight = 0.8,
                      comp.dist = list("multinom", "multinom"),
                      comp.param = list(list("size" = 1, "prob" = c(0.2,0.3,0.5)),
                                        list("size" = 1, "prob" = c(0.1,0.6,0.3))))
data1 <- getmixtData(mixt1)
mixt2 <- twoComp_mixt(n = 2000, weight = 0.3,
                      comp.dist = list("multinom", "multinom"),
                      comp.param = list(list("size" = 1, "prob" = c(0.2,0.3,0.5)),
                                        list("size" = 1, "prob" = c(0.7,0.1,0.2))))
data2 <- getmixtData(mixt2)
## Define the admixture models:
admixMod1 <- admix_model(knownComp_dist = mixt1$comp.dist[[2]],
                         knownComp_param = mixt1$comp.param[[2]])
admixMod2 <- admix_model(knownComp_dist = mixt2$comp.dist[[2]],
                         knownComp_param = mixt2$comp.param[[2]])
## Estimate the mixture weights of the two admixture models (provide only hat(theta)_n):
estim_IBM(samples = list(data1,data2), admixMod = list(admixMod1,admixMod2),
          compute_var = TRUE)
#> Warning:  IBM estimators of two unknown proportions are reliable only if the two
#>     corresponding unknown component distributions have been tested equal (see 'admix_test()').
#>     Furthermore, when both the known and unknown component distributions of the mixture
#>     models are identical, the IBM approach provides an estimation of the ratio of the
#>     actual mixing weights rather than an estimation of the unknown weights themselves.
#> $n_populations
#> [1] 2
#> 
#> $population_sizes
#> [1] 1500 2000
#> 
#> $admixture_models
#> $admixture_models[[1]]
#> Call:admix_model(knownComp_dist = mixt1$comp.dist[[2]], knownComp_param = mixt1$comp.param[[2]])
#> 
#> Known component distribution:  multinom 
#> Value of the known component parameters:
#> $size
#> [1] 1
#> 
#> $prob
#> [1] 0.1 0.6 0.3
#> 
#> 
#> 
#> $admixture_models[[2]]
#> Call:admix_model(knownComp_dist = mixt2$comp.dist[[2]], knownComp_param = mixt2$comp.param[[2]])
#> 
#> Known component distribution:  multinom 
#> Value of the known component parameters:
#> $size
#> [1] 1
#> 
#> $prob
#> [1] 0.7 0.1 0.2
#> 
#> 
#> 
#> 
#> $estimation_method
#> [1] "Inversion Best Matching (IBM)"
#> 
#> $estimated_mixing_weights
#> [1] 0.7187107 0.3061926
#> 
#> $variance_est_p1
#> [1] 0.005796882
#> 
#> $variance_est_p2
#> [1] 0.0005513328
#> 
#> $equal.knownComp
#> [1] FALSE
#> 
#> $p.X.fixed
#> NULL
#> 
#> $integ.supp
#> [1] 1 2 3
#> 
#> $call
#> estim_IBM(samples = list(data1, data2), admixMod = list(admixMod1, 
#>     admixMod2), compute_var = TRUE)
#> 
#> attr(,"class")
#> [1] "estim_IBM"   "admix_estim"