Variance of estimators in an admixture model with symmetric unknown density.

Semiparametric estimation of the variance of the estimators related to the mixture weight p and the location shift parameter mu, considering the admixture model with probability density function l: l(x) = p*f(x-mu) + (1-p)*g(x), x in R, where g is the known component of the two-component mixture, p is the unknown proportion, f is the unknown component density and mu is the location shift. See 'Details' below for more information.

Usage

BVdk_varCov_estimators(estim, data, admixMod)

Arguments

estim

An object of class 'estim_BVdk', containing the estimators of unknown quantities in the admixture model.

data

The observed sample under study.

admixMod

An object of class 'admix_model', containing useful information about distributions and parameters.

Value

A list containing 1) the variance-covariance matrix of the estimators (assessed at the specific time points 'u' and 'v' such that u = v = mean(data)); 2) the variance of the estimator of the unknown mixture weight; 3) the variance of the estimator of the location shift parameter; 4) the variance of the estimator of the unknown component cumulative distribution function at points 'u' and 'v'.

Details

See formulas pp.28--30 in Appendix of Bordes, L. and Vandekerkhove, P. (2010).

References

Bordes L, Vandekerkhove P (2010). “Semiparametric two-component mixture model with a known component: An asymptotically normal estimator.” Mathematical Methods of Statistics, 19(1), 22--41. doi:10.3103/S1066530710010023 .

Author

Xavier Milhaud xavier.milhaud.research@gmail.com

Examples

## Simulate mixture data:
mixt1 <- twoComp_mixt(n = 200, weight = 0.4,
                      comp.dist = list("norm", "norm"),
                      comp.param = list(list("mean" = -2, "sd" = 0.5),
                                        list("mean" = 0, "sd" = 1)))
data1 <- getmixtData(mixt1)
## Define the admixture model:
admixMod <- admix_model(knownComp_dist = mixt1$comp.dist[[2]],
                        knownComp_param = mixt1$comp.param[[2]])

## Perform the estimation of parameters in real-life:
estim <- estim_BVdk(samples = data1, admixMod = admixMod, method = 'L-BFGS-B')
#> Warning: Estimation by 'BVdk' assumes the unknown component distribution
#>   to be a standard Gaussian distribution.
BVdk_varCov_estimators(estim = estim, data = data1, admixMod = admixMod)
#> $varCov.mat
#>             [,1]          [,2]          [,3]
#> [1,] 0.002415413  0.0023685087  0.0010002504
#> [2,] 0.002368509  0.0097004497 -0.0005787659
#> [3,] 0.001000250 -0.0005787659 -0.0002626145
#> 
#> $var.estim_prop
#> [1] 0.002415413
#> 
#> $var.estim_location
#> [1] 0.00970045
#> 
#> $var.estim_cdf
#> [1] -0.0002626145
#>