Simulated distribution of the contrast using IBM

Tabulate the distribution related to the inner convergence part of the contrast, by simulating trajectories of Gaussian processes and applying some transformations. Useful to perform the test hypothesis, by retrieving the (1-alpha)-quantile of interest. See 'Details' below and the cited paper therein for further information.

Usage

IBM_tabul_stochasticInteg(
  n.sim = 200,
  n.varCovMat = 100,
  samples,
  admixMod,
  min_size = NULL,
  parallel = FALSE,
  n_cpu = 2
)

Arguments

n.sim

Number of trajectories of simulated gaussian processes (number of random draws for tabulation).

n.varCovMat

Number of time points on which gaussian processes are simulated.

samples

(list) A list of the two samples under study.

admixMod

(list) A list of two objects of class 'admix_model', containing useful information about distributions and parameters.

min_size

(optional, NULL by default) In the k-sample case, useful to provide the minimal size among all samples (needed to take into account the correction factor for variance-covariance assessment). Otherwise, useless.

parallel

(default to FALSE) Boolean to indicate whether parallel computations are performed (speed-up the tabulation).

n_cpu

(default to 2) Number of cores used for computations when parallelizing.

Value

A list with four elements, containing: 1) random draws of the quantity 'sample size times the empirical contrast', as defined in the IBM approach (see 'Details' above); 2) the estimated unknown component weights for the two admixture models; 3) the value of the quantity 'sample size times the empirical contrast'; 4) the sequence of points in the support that were used to evaluate the variance-covariance matrix of empirical processes.

References

Milhaud X, Pommeret D, Salhi Y, Vandekerkhove P (2024). “Two-sample contamination model test.” Bernoulli, 30(1), 170--197. doi:10.3150/23-BEJ1593 .

Author

Xavier Milhaud xavier.milhaud.research@gmail.com

Examples

# \donttest{
## Simulate mixture data:
mixt1 <- twoComp_mixt(n = 1200, weight = 0.4,
                      comp.dist = list("norm", "norm"),
                      comp.param = list(list("mean" = -2, "sd" = 0.5),
                                        list("mean" = 0, "sd" = 1)))
mixt2 <- twoComp_mixt(n = 1000, weight = 0.7,
                      comp.dist = list("norm", "norm"),
                      comp.param = list(list("mean" = -2, "sd" = 0.5),
                                        list("mean" = 1, "sd" = 1)))
data1 <- getmixtData(mixt1)
data2 <- getmixtData(mixt2)
## Define the admixture models:
admixMod1 <- admix_model(knownComp_dist = mixt1$comp.dist[[2]],
                         knownComp_param = mixt1$comp.param[[2]])
admixMod2 <- admix_model(knownComp_dist = mixt2$comp.dist[[2]],
                         knownComp_param = mixt2$comp.param[[2]])
IBM_tabul_stochasticInteg(n.sim = 2, n.varCovMat = 20, samples = list(data1, data2),
                          admixMod = list(admixMod1, admixMod2), min_size = NULL,
                          parallel = FALSE, n_cpu = 2)
#> $U_sim
#> [1] 0.6780141 0.1884194
#> 
#> $estimator
#> Call:estim_IBM(samples = samples, admixMod = admixMod, n.integ = 1000)
#> 
#> Estimated mixing proportion (of the unknown component) in the 1st sample:  0.3399946 
#> Estimated mixing proportion (of the unknown component) in the 2nd sample:  0.6540184 
#> 
#> Fixed value for the mixing weight in the 1st sample (case of identical known components in the 2 samples): 
#> 
#> 
#> Support of integration: -3.718936 -3.711886 -3.707694 -3.707643 -3.703424 -3.675233 -3.672924 -3.663755 -3.661976 -3.653663 -3.648718 -3.647641 -3.645167 -3.636859 -3.630708 -3.607828 -3.604966 -3.604073 -3.599054 -3.589697 -3.581929 -3.567491 -3.566806 -3.553931 -3.545407 -3.541853 -3.538392 -3.534202 -3.521839 -3.521832 -3.518533 -3.509769 -3.499132 -3.489261 -3.485159 -3.467691 -3.445911 -3.443413 -3.440146 -3.439103 -3.432303 -3.416664 -3.400446 -3.398392 -3.391027 -3.390518 -3.386411 -3.381423 -3.372104 -3.371469 -3.369402 -3.356564 -3.349916 -3.342126 -3.335661 -3.333947 -3.332676 -3.328376 -3.318463 -3.298475 -3.29049 -3.288452 -3.279041 -3.275484 -3.27163 -3.270018 -3.266387 -3.261563 -3.259987 -3.258502 -3.249964 -3.235022 -3.234307 -3.231229 -3.228514 -3.225452 -3.224002 -3.201461 -3.200504 -3.1901 -3.186074 -3.15609 -3.153962 -3.153012 -3.152041 -3.146025 -3.145633 -3.141885 -3.137771 -3.127258 -3.123401 -3.112385 -3.11015 -3.106712 -3.099159 -3.095786 -3.091902 -3.090447 -3.077365 -3.076563 -3.074931 -3.067837 -3.064972 -3.044833 -3.032952 -3.029237 -3.016347 -3.01103 -3.001539 -2.998067 -2.993883 -2.962995 -2.956988 -2.948315 -2.94387 -2.935189 -2.920738 -2.918677 -2.913918 -2.905683 -2.903949 -2.900483 -2.891537 -2.887711 -2.853934 -2.84579 -2.843312 -2.832577 -2.827766 -2.820552 -2.800513 -2.791949 -2.789385 -2.786929 -2.730096 -2.72732 -2.724208 -2.708773 -2.70466 -2.701614 -2.701013 -2.699926 -2.67855 -2.673835 -2.671916 -2.657989 -2.647247 -2.6462 -2.64329 -2.635632 -2.629464 -2.628475 -2.62572 -2.619031 -2.617728 -2.616577 -2.611294 -2.605544 -2.593981 -2.591057 -2.581842 -2.581425 -2.575501 -2.572063 -2.565763 -2.560214 -2.553234 -2.552384 -2.547804 -2.537258 -2.534489 -2.523738 -2.49687 -2.496307 -2.491489 -2.484717 -2.477673 -2.475984 -2.465373 -2.456479 -2.450641 -2.446381 -2.432009 -2.405183 -2.387733 -2.383829 -2.375309 -2.369956 -2.369744 -2.344261 -2.339841 -2.327132 -2.325873 -2.325436 -2.319876 -2.299978 -2.299195 -2.296152 -2.273497 -2.269508 -2.256751 -2.248654 -2.239555 -2.223795 -2.221671 -2.211323 -2.188534 -2.175044 -2.1654 -2.147864 -2.140959 -2.121672 -2.107123 -2.102985 -2.099581 -2.091422 -2.090739 -2.089704 -2.085168 -2.07455 -2.059951 -2.045074 -2.044732 -2.035817 -2.018445 -2.0058 -1.978847 -1.974802 -1.971769 -1.96935 -1.947175 -1.943729 -1.939968 -1.935429 -1.912726 -1.907659 -1.90416 -1.876195 -1.869456 -1.859934 -1.842557 -1.83819 -1.817137 -1.795409 -1.794845 -1.792423 -1.775699 -1.756301 -1.755706 -1.747599 -1.742337 -1.731381 -1.730339 -1.724385 -1.710005 -1.709698 -1.70871 -1.692784 -1.684561 -1.679341 -1.662824 -1.652578 -1.652166 -1.647375 -1.645679 -1.639521 -1.621574 -1.615312 -1.613658 -1.595868 -1.592404 -1.589702 -1.582617 -1.581138 -1.579124 -1.576764 -1.564033 -1.556803 -1.549589 -1.548033 -1.525002 -1.506799 -1.479912 -1.479181 -1.459491 -1.458569 -1.445648 -1.437331 -1.437217 -1.430589 -1.413246 -1.406902 -1.40231 -1.392944 -1.383794 -1.379966 -1.377149 -1.374331 -1.369209 -1.366625 -1.358999 -1.358165 -1.355058 -1.339533 -1.336169 -1.306903 -1.305039 -1.304754 -1.303849 -1.296355 -1.293129 -1.273838 -1.262603 -1.253011 -1.242019 -1.236573 -1.231836 -1.219366 -1.216349 -1.210807 -1.204963 -1.202111 -1.199179 -1.176435 -1.17475 -1.148877 -1.141197 -1.140034 -1.127857 -1.115221 -1.112966 -1.106262 -1.10596 -1.105515 -1.104068 -1.10104 -1.099814 -1.099789 -1.096366 -1.088507 -1.085151 -1.083091 -1.082002 -1.077944 -1.063926 -1.045223 -1.032325 -1.03097 -1.022833 -1.022795 -1.02265 -1.01896 -1.007251 -1.006357 -1.001801 -0.9966549 -0.9956455 -0.9947595 -0.9839213 -0.9801741 -0.9671475 -0.9636953 -0.9625033 -0.951964 -0.9515894 -0.9208884 -0.9182842 -0.9119214 -0.9064292 -0.89998 -0.8983526 -0.8882798 -0.876194 -0.87566 -0.8755592 -0.8723359 -0.8652886 -0.864997 -0.8636734 -0.8571117 -0.8537619 -0.8425081 -0.8344453 -0.8292193 -0.8274212 -0.8273764 -0.8133745 -0.8103473 -0.7999329 -0.7993413 -0.7613934 -0.7577081 -0.7361141 -0.729766 -0.7154972 -0.7045786 -0.7027841 -0.6982138 -0.6932669 -0.6841674 -0.6807579 -0.6782603 -0.670876 -0.6669079 -0.6609471 -0.6548865 -0.6408242 -0.5980511 -0.5903472 -0.5871931 -0.5781154 -0.5763372 -0.5752889 -0.5443352 -0.5416898 -0.5329333 -0.5258893 -0.5165408 -0.5131574 -0.5131336 -0.5103303 -0.5071973 -0.4997238 -0.4992919 -0.4914229 -0.4910991 -0.4902658 -0.4874378 -0.4857229 -0.4848016 -0.4753327 -0.4702436 -0.4664166 -0.464548 -0.4620615 -0.4595329 -0.4527106 -0.4507605 -0.4450769 -0.4189255 -0.4135511 -0.413264 -0.4126106 -0.410141 -0.4058077 -0.40411 -0.385638 -0.3818415 -0.3750479 -0.3733483 -0.3370259 -0.3328311 -0.3313191 -0.3241314 -0.323535 -0.3085611 -0.298971 -0.2857013 -0.2854307 -0.2699817 -0.2679653 -0.264962 -0.2643317 -0.2561175 -0.2443485 -0.2407041 -0.2306319 -0.2290345 -0.2228553 -0.2223782 -0.2033125 -0.1978261 -0.1932736 -0.1898212 -0.1889459 -0.1756488 -0.1720723 -0.1650889 -0.1602694 -0.1441687 -0.1307313 -0.1266115 -0.1230487 -0.1181121 -0.1179779 -0.1076378 -0.103798 -0.08624939 -0.08516943 -0.08101837 -0.07565025 -0.07259462 -0.06726838 -0.06040409 -0.05407813 -0.05405279 -0.05394349 -0.04673739 -0.0444489 -0.0219703 -0.003138096 0.02804856 0.04962131 0.05088438 0.05394179 0.05540504 0.05548078 0.07991045 0.08976471 0.0947002 0.1021158 0.1091718 0.1174785 0.1198795 0.1290332 0.1304345 0.1446577 0.1460577 0.1508277 0.1560697 0.163823 0.1647615 0.1811721 0.181409 0.1844259 0.1951902 0.1964377 0.1968077 0.2026306 0.2054897 0.2138266 0.2286041 0.2323221 0.234837 0.2370385 0.2403625 0.2409025 0.2640548 0.2686206 0.2746982 0.2804892 0.2861718 0.2885591 0.3034356 0.3055581 0.3328311 0.3486921 0.3609003 0.3703923 0.3819496 0.3907255 0.3979069 0.4002703 0.4047408 0.4053747 0.4065297 0.4106235 0.4222624 0.4250098 0.4307255 0.447759 0.4719848 0.480449 0.5242527 0.5331034 0.5351451 0.5400246 0.54364 0.5441067 0.5512637 0.5657429 0.5787295 0.5883593 0.5975502 0.598312 0.6096082 0.6203081 0.6286701 0.6299785 0.630391 0.6332488 0.6347126 0.6456841 0.6493077 0.6586818 0.684807 0.6858178 0.691628 0.695315 0.6990113 0.7034464 0.7160349 0.7197101 0.7303959 0.7507376 0.7508127 0.7534891 0.7740286 0.7946137 0.7959866 0.7998902 0.8020948 0.8028339 0.8164931 0.8174533 0.8253741 0.8315637 0.8401364 0.8429976 0.8498004 0.8504966 0.8614331 0.8825561 0.8833421 0.8850068 0.8879587 0.888849 0.8939513 0.9016948 0.9153985 0.9183398 0.9199935 0.9269394 0.9669139 0.9683377 0.9917345 0.9942857 1.001359 1.005573 1.009905 1.022381 1.035166 1.035427 1.039186 1.04685 1.046902 1.071195 1.075223 1.082209 1.088578 1.090046 1.091158 1.091431 1.100526 1.12284 1.127498 1.151367 1.166551 1.169541 1.178324 1.181441 1.187376 1.191103 1.198 1.198353 1.206125 1.227105 1.235023 1.257157 1.260262 1.261565 1.262485 1.270412 1.270658 1.289652 1.296374 1.301581 1.312144 1.312456 1.313307 1.320362 1.350269 1.352472 1.358685 1.358828 1.365584 1.379074 1.386662 1.406182 1.407763 1.407893 1.42103 1.442001 1.444112 1.44625 1.462226 1.467294 1.484831 1.486048 1.490546 1.496004 1.507135 1.513881 1.517099 1.529593 1.537066 1.562651 1.573763 1.591599 1.628041 1.631564 1.645654 1.649351 1.660753 1.664263 1.664417 1.666935 1.685341 1.708377 1.714972 1.753175 1.753296 1.754729 1.755699 1.765902 1.775414 1.78984 1.82342 1.832509 1.847243 1.848247 1.853069 1.861787 1.862352 1.862665 1.863525 1.863716 1.866094 1.877312 1.892147 1.89765 1.903165 1.906162 1.922705 1.924584 1.924884 1.929162 1.940881 1.941278 1.941554 1.944392 1.952316 1.959926 1.981366 1.985322 1.985508 1.990274 1.993737 1.999201 2.004348 2.011382 2.013745 2.017634 2.033948 2.035931 2.040173 2.043851 2.049376 2.055655 2.060421 2.069277 2.073414 2.080733 2.088672 2.122566 2.125633 2.131443 2.139868 2.140344 2.143715 2.144198 2.147708 2.167583 2.170323 2.203304 2.215361 2.244109 2.24426 2.272919 2.274445 2.275531 2.278468 2.279306 2.28758 2.309575 2.313158 2.35251 2.365847 2.367792 2.375769 2.382083 2.390373 2.407362 2.408589 2.423072 2.43741 2.441783 2.446746 2.450602 2.451464 2.452898 2.455935 2.462695 2.463396 2.474769 2.479769 2.48913 2.50506 2.512366 2.51823 2.526418 2.529727 2.548331 2.551841 2.552555 2.556587 2.569071 2.573293 2.576276 2.57889 2.584872 2.585652 2.588989 2.603394 2.607105 2.612814 2.61296 2.613204 2.614714 2.623769 2.626241 2.626585 2.644599 2.658118 2.660158 2.670354 2.672369 2.685467 2.691944 2.702641 2.715334 2.717713 2.724673 2.726868 2.73677 2.744303 2.751361 2.751509 2.756932 2.773717 2.775466 2.775901 2.778493 2.797027 2.799282 2.812266 2.812756 2.814728 2.818184 2.823996 2.849836 2.855806 2.861297 2.861578 2.867141 2.869629 2.87601 2.881275 2.881702 2.887885 2.88988 2.894922 2.897351 2.91912 2.920642 2.927845 2.936846 2.965002 2.966539 2.969174 2.980338 2.982906 2.983044 2.986776 3.007732 3.019786 3.028548 3.039835 3.040052 3.043578 3.044241 3.050676 3.05242 3.061319 3.068506 3.074121 3.112717 3.112983 3.119863 3.121714 3.125754 3.159653 3.163617 3.170511 3.181277 3.183276 3.189725 3.196318 3.202441 3.204953 3.210704 3.215742 3.217996 3.21801 3.227694 3.230459 3.242129 3.248357 3.248859 3.253356 3.258593 3.271088 3.27628 3.282946 3.283297 3.287585 3.294454 3.295308 3.3042 3.313856 3.313996 3.314468 3.325128 3.332364 3.345955 3.346958 3.348577 3.36007 3.36633 3.369031 3.381491 3.384861 3.385743 3.393913 3.396154 3.400556 3.401803 3.404189 3.420862 3.423985 3.426414 3.434002 3.439335 3.460576 3.473116 3.481054 3.483041 3.483201 3.486265 3.493463 3.495418 3.495733 3.499719 3.506537 3.511887 3.516987 3.517849 3.518004 3.520933 3.534571 3.542994 3.546178 3.546829 3.552873 3.552949 3.557766 3.559605 3.574051 3.600221 3.600814 3.604232 3.616129 3.626555 3.634604 3.636517 3.641493 3.652066 3.657713 3.659204 3.667906 3.674374 3.675251 3.678447 3.680506 3.683496 3.687568 3.694949 3.69522 3.698496 3.702835 
#> 
#> 
#> $contrast_value
#> [1] 0.1315202
#> 
#> $integ.points
#>  [1] -3.7189356 -3.3283162 -2.9376967 -2.5470772 -2.1564577 -1.7658382
#>  [7] -1.3752187 -0.9845992 -0.5939797 -0.2033602  0.1872593  0.5778788
#> [13]  0.9684982  1.3591177  1.7497372  2.1403567  2.5309762  2.9215957
#> [19]  3.3122152  3.7028347
#> 
# }