Estimate the unknown parameters of the admixture model(s)

Estimate the component weights, the location shift parameter (in case of a symmetric unknown component density), and the unknown component distribution using different estimation techniques. We remind that the i-th admixture model has probability density function (pdf) l_i such that: l_i = p_i * f_i + (1-p_i) * g_i, where g_i is the known component density. The unknown quantities p_i and f_i then have to be estimated.

Usage

admix_estim(
  samples,
  admixMod,
  est.method = c("BVdk", "PS", "IBM"),
  sym.f = FALSE
)

Arguments

samples

(list) A list of the K (K>0) samples to be studied, all following admixture distributions.

admixMod

(list) A list of objects of class 'admix_model', containing useful information about distributions and parameters.

est.method

The estimation method to be applied. Can be one of 'BVdk' (Bordes and Vandekerkhove estimator), 'PS' (Patra and Sen estimator), or 'IBM' (Inversion Best-Matching approach). The same estimation method is performed on each sample. Important note: estimation by 'IBM' is unbiased only under H0, meaning that choosing this method requires to perform previously the test hypothesis between the pairs of samples. For further details, see section 'Details' below.

sym.f

A boolean indicating whether the unknown component densities are assumed to be symmetric or not.

Value

A list containing the estimated weight of every unknown component distribution among admixture samples.

Details

For further details on the different estimation techniques, see references below i) Patra and Sen estimator ; ii) BVdk estimator ; iii) IBM approach.

References

PatraSen2016admix BordesDelmasVandekerkhove2006admix BordesVandekerkhove2010admix MilhaudPommeretSalhiVandekerkhove2024aadmix

Author

Xavier Milhaud xavier.milhaud.research@gmail.com

Examples

## Simulate mixture data:
mixt1 <- twoComp_mixt(n = 450, weight = 0.4,
                      comp.dist = list("norm", "norm"),
                      comp.param = list(list("mean" = -2, "sd" = 0.5),
                                        list("mean" = 0, "sd" = 1)))
mixt2 <- twoComp_mixt(n = 380, weight = 0.7,
                      comp.dist = list("norm", "norm"),
                      comp.param = list(list("mean" = -2, "sd" = 0.5),
                                        list("mean" = 1, "sd" = 1)))
mixt3 <- twoComp_mixt(n = 400, weight = 0.8,
                      comp.dist = list("norm", "exp"),
                      comp.param = list(list("mean" = 3, "sd" = 1),
                                        list("rate" = 1)))
data1 <- getmixtData(mixt1)
data2 <- getmixtData(mixt2)
data3 <- getmixtData(mixt3)

## Define the admixture models:
admixMod1 <- admix_model(knownComp_dist = mixt1$comp.dist[[2]],
                         knownComp_param = mixt1$comp.param[[2]])
admixMod2 <- admix_model(knownComp_dist = mixt2$comp.dist[[2]],
                         knownComp_param = mixt2$comp.param[[2]])
admixMod3 <- admix_model(knownComp_dist = mixt3$comp.dist[[2]],
                         knownComp_param = mixt3$comp.param[[2]])

admix_estim(samples = list(data1,data2,data3),
            admixMod = list(admixMod1,admixMod2,admixMod3),
            est.method = 'BVdk', sym.f = TRUE)
#> Call:
#> admix_estim(samples = list(data1, data2, data3), admixMod = list(admixMod1, 
#>     admixMod2, admixMod3), est.method = "BVdk", sym.f = TRUE)
#> 
#> Estimated mixing weight of the unknown component distribution in Sample 1: 0.4
#> Estimated mixing weight of the unknown component distribution in Sample 2: 0.7
#> Estimated mixing weight of the unknown component distribution in Sample 3: 0.83
#> 
#> Estimated location parameters of the unknown component distribution in Sample 1: -1.98
#> Estimated location parameters of the unknown component distribution in Sample 2: -1.97
#> Estimated location parameters of the unknown component distribution in Sample 3: 2.92
admix_estim(samples = list(data1,data2,data3),
            admixMod = list(admixMod1,admixMod2,admixMod3),
            est.method = 'PS')
#> Call:
#> admix_estim(samples = list(data1, data2, data3), admixMod = list(admixMod1, 
#>     admixMod2, admixMod3), est.method = "PS")
#> 
#> Estimated mixing weight of the unknown component distribution in Sample 1: 0.38
#> Estimated mixing weight of the unknown component distribution in Sample 2: 0.69
#> Estimated mixing weight of the unknown component distribution in Sample 3: 0.75
#> 
admix_estim(samples = list(data1,data2,data3),
            admixMod = list(admixMod1,admixMod2,admixMod3),
            est.method = 'IBM')
#>     (IBM) Estimators of two unknown proportions are reliable only if
#>     the two corresponding unknown component distributions have been tested equal
#>     (see function 'admix_test').
#> 
#>     Furthermore, when both the known and unknown component distributions of the mixture
#>     models are identical, the IBM approach provides an estimation of the ratio of the
#>     actual mixing weights rather than an estimation of the unknown weights themselves.
#> Call:
#> admix_estim(samples = list(data1, data2, data3), admixMod = list(admixMod1, 
#>     admixMod2, admixMod3), est.method = "IBM")
#> 
#> Pairwise estimation performed (between samples listed as column names): 
#>      Samples 1 and 2/ Samples 1 and 3/
#> [1,]             0.44                5
#> [2,]             0.75                5
#> First row provides estimated mixing weight in the first of the two samples listed,
#>     and second row the one in the second sample.
#>