Estimates weights of unknown components from 2 admixtures using IBM

Estimation of the component weights from the Inversion - Best Matching (IBM) method, related to two admixture models with respective probability density function (pdf) l1 and l2, such that: l1 = p1*f1 + (1-p1)g1 and l2 = p2f2 + (1-p2)*g2, where g1 and g2 are the known component densities. For further details about IBM approach, see 'Details' below.

Usage

estim_IBM(samples, admixMod, n.integ = 1000)

Arguments

samples

(List) List of the two considered samples.

admixMod

(List) List of objects of class 'admix_model', one for each sample.

n.integ

Number of data points generated for the distribution on which to integrate.

Value

A list with the two estimates of the component weights for each of the admixture model, plus that of the theoretical model if specified.

References

Milhaud X, Pommeret D, Salhi Y, Vandekerkhove P (2024). “Two-sample contamination model test.” Bernoulli, 30(1), 170--197. doi:10.3150/23-BEJ1593 .

Author

Xavier Milhaud xavier.milhaud.research@gmail.com

Examples

## Simulate mixture data:
mixt1 <- twoComp_mixt(n = 1500, weight = 0.5,
                      comp.dist = list("norm", "norm"),
                      comp.param = list(list("mean" = 3, "sd" = 0.5),
                                        list("mean" = 0, "sd" = 1)))
data1 <- getmixtData(mixt1)
mixt2 <- twoComp_mixt(n = 2000, weight = 0.7,
                      comp.dist = list("norm", "norm"),
                      comp.param = list(list("mean" = 3, "sd" = 0.5),
                                        list("mean" = 5, "sd" = 2)))
data2 <- getmixtData(mixt2)

## Define the admixture models:
admixMod1 <- admix_model(knownComp_dist = mixt1$comp.dist[[2]],
                         knownComp_param = mixt1$comp.param[[2]])
admixMod2 <- admix_model(knownComp_dist = mixt2$comp.dist[[2]],
                         knownComp_param = mixt2$comp.param[[2]])

## Estimate the mixture weights of the two admixture models (provide only hat(theta)_n):
estim_IBM(samples = list(data1,data2),
          admixMod = list(admixMod1,admixMod2), n.integ = 1000)
#> Call:estim_IBM(samples = list(data1, data2), admixMod = list(admixMod1, 
#>     admixMod2), n.integ = 1000)
#> 
#> Estimated mixing proportion (of the unknown component) in the 1st sample:  0.5016692 
#> Estimated mixing proportion (of the unknown component) in the 2nd sample:  0.7008618 
#> 
#> Fixed value for the mixing weight in the 1st sample (case of identical known components in the 2 samples): 
#> 
#> 
#> Support of integration: -2.951133 -2.935747 -2.921151 -2.890041 -2.889465 -2.851137 -2.844893 -2.80197 -2.787619 -2.710968 -2.695534 -2.6783 -2.653893 -2.552877 -2.552177 -2.54101 -2.526621 -2.480118 -2.471756 -2.470035 -2.46415 -2.438351 -2.433668 -2.417662 -2.408153 -2.406664 -2.397827 -2.383797 -2.382519 -2.353424 -2.352745 -2.339048 -2.3362 -2.336127 -2.332926 -2.325494 -2.304568 -2.302165 -2.296096 -2.295173 -2.284263 -2.222674 -2.125444 -2.121272 -2.101405 -2.090125 -2.078117 -2.063091 -2.048056 -2.010865 -2.010292 -2.009709 -2.008516 -1.994815 -1.99092 -1.96895 -1.964272 -1.95972 -1.955146 -1.952522 -1.9483 -1.935761 -1.92845 -1.881154 -1.877171 -1.869223 -1.865924 -1.817684 -1.80211 -1.766912 -1.766129 -1.762259 -1.746603 -1.694352 -1.694192 -1.680716 -1.659183 -1.650819 -1.649326 -1.642317 -1.624655 -1.616201 -1.591891 -1.582157 -1.580775 -1.564937 -1.557406 -1.545364 -1.541067 -1.532061 -1.525789 -1.51898 -1.467141 -1.454712 -1.452878 -1.426183 -1.41451 -1.406594 -1.400617 -1.398562 -1.384264 -1.378726 -1.327407 -1.270627 -1.270443 -1.263487 -1.209113 -1.195488 -1.192143 -1.187746 -1.169199 -1.135743 -1.132411 -1.118607 -1.089883 -1.087651 -1.080845 -1.048448 -1.048382 -1.045507 -1.035334 -1.012417 -1.008034 -1.005086 -0.9787743 -0.9629727 -0.9445054 -0.9359439 -0.8829082 -0.8713857 -0.8404673 -0.8057453 -0.803266 -0.8011673 -0.7878297 -0.7855432 -0.7608315 -0.7571117 -0.7566666 -0.7435748 -0.7363124 -0.7315557 -0.7301408 -0.7276976 -0.7219632 -0.7119078 -0.680863 -0.675528 -0.6285124 -0.5914088 -0.5799705 -0.5572773 -0.5570696 -0.5360852 -0.5311359 -0.5191334 -0.47059 -0.4380624 -0.4371511 -0.3948237 -0.3435087 -0.3332691 -0.30266 -0.2893659 -0.2795929 -0.2764862 -0.2350472 -0.2348112 -0.2149009 -0.210725 -0.2057196 -0.1874462 -0.1837902 -0.182477 -0.1736318 -0.155549 -0.1517332 -0.1496524 -0.1419284 -0.1117871 -0.09621403 -0.08612575 -0.06689552 -0.05611482 -0.05214409 -0.04113781 -0.03841473 -0.03353392 -0.02720039 -0.02071223 -0.01695829 -0.01633507 0.01590548 0.02812985 0.0324389 0.0366007 0.04268435 0.06281002 0.06458636 0.07702545 0.08089357 0.08217193 0.08987039 0.09057279 0.1011305 0.1022708 0.121769 0.1230327 0.162633 0.1785709 0.2282471 0.2339297 0.2365852 0.2756911 0.2964489 0.3092379 0.3171507 0.3247187 0.3351999 0.3535849 0.3596147 0.3677425 0.3691865 0.3698629 0.4004692 0.4021863 0.4174378 0.4436052 0.452892 0.4638723 0.5001913 0.5006903 0.5017052 0.528028 0.5522096 0.5770161 0.6012081 0.620047 0.6329424 0.6467877 0.6589875 0.6923864 0.696903 0.6972712 0.7014917 0.7358255 0.7637364 0.7868926 0.7948318 0.7951495 0.8083998 0.822333 0.865115 0.8781232 0.8962111 0.8975055 0.8991311 0.9045978 0.9190482 0.9286143 0.9315397 0.9432549 1.00538 1.00947 1.011358 1.023222 1.033472 1.060194 1.111946 1.165028 1.18928 1.225559 1.226636 1.249086 1.257712 1.258652 1.268763 1.268998 1.291252 1.292338 1.311537 1.319734 1.336816 1.364726 1.386557 1.399205 1.433213 1.444602 1.466677 1.483619 1.494582 1.499794 1.504472 1.506696 1.532335 1.537996 1.540836 1.541158 1.557645 1.565056 1.567197 1.5799 1.622624 1.663189 1.679858 1.702219 1.720372 1.733083 1.74563 1.752249 1.759823 1.763351 1.769783 1.82699 1.832498 1.839431 1.880333 1.902511 1.905877 1.915193 1.943211 1.951684 1.959667 1.962152 1.980675 1.995262 2.005853 2.007535 2.018999 2.055525 2.077011 2.12959 2.131619 2.142709 2.145646 2.148152 2.181614 2.207594 2.223751 2.225589 2.24268 2.246624 2.324628 2.332695 2.353106 2.446945 2.449614 2.462978 2.47068 2.478382 2.493694 2.508923 2.529013 2.533982 2.542157 2.54896 2.56519 2.569263 2.583282 2.583917 2.588204 2.589722 2.598718 2.599785 2.608717 2.608826 2.703749 2.709239 2.720907 2.748869 2.791079 2.791493 2.798256 2.826523 2.826615 2.827175 2.837636 2.892199 2.917165 2.950657 2.956202 2.959968 2.969608 2.970772 2.989044 3.002397 3.009391 3.022736 3.029892 3.041159 3.045055 3.060089 3.085978 3.095991 3.121729 3.134681 3.170558 3.176704 3.190253 3.200489 3.20996 3.215556 3.222505 3.238244 3.249404 3.291201 3.312537 3.35761 3.363522 3.386908 3.426923 3.432805 3.442983 3.483283 3.537751 3.542244 3.550989 3.554059 3.565702 3.566018 3.572372 3.582012 3.585058 3.587732 3.598035 3.616837 3.628657 3.648672 3.66622 3.687476 3.689175 3.701724 3.72039 3.733447 3.74609 3.783328 3.787733 3.82932 3.839401 3.855974 3.874635 3.875784 3.892081 3.92997 3.933924 3.941388 3.967397 3.970178 3.971376 3.980851 3.990305 3.992363 4.006956 4.026864 4.02714 4.028677 4.078712 4.096898 4.099403 4.100503 4.144824 4.160472 4.16194 4.168071 4.20974 4.221477 4.222989 4.290643 4.291327 4.308769 4.336476 4.347981 4.361696 4.397881 4.430402 4.435981 4.449435 4.471947 4.473847 4.479299 4.487104 4.492347 4.518863 4.545621 4.565284 4.565678 4.575404 4.579662 4.607319 4.610275 4.638382 4.651068 4.703559 4.714913 4.716298 4.731936 4.738257 4.739938 4.743917 4.764623 4.786991 4.79386 4.829358 4.841536 4.852878 4.854419 4.855964 4.859721 4.871868 4.875623 4.932337 4.9611 4.971341 4.976474 4.981835 5.009176 5.031461 5.03356 5.056303 5.087499 5.090237 5.115558 5.116366 5.127521 5.128062 5.152318 5.1606 5.186263 5.195502 5.205489 5.214764 5.218837 5.241086 5.254642 5.277154 5.285024 5.30745 5.31983 5.350503 5.350856 5.365611 5.384926 5.395806 5.409648 5.424828 5.433545 5.466479 5.469031 5.482412 5.490867 5.498966 5.509001 5.509056 5.515569 5.540672 5.58287 5.59064 5.591569 5.594638 5.620439 5.628619 5.663499 5.671422 5.671791 5.684323 5.696457 5.703436 5.711944 5.742007 5.751073 5.752565 5.754993 5.779156 5.783812 5.792886 5.805123 5.849709 5.850786 5.864575 5.875852 5.923988 5.934491 5.957326 5.957734 5.969696 5.978927 5.991747 6.018115 6.02423 6.058912 6.072818 6.080392 6.086618 6.087649 6.146357 6.147506 6.154224 6.157968 6.207615 6.22902 6.229268 6.235066 6.237583 6.243912 6.248711 6.259944 6.26074 6.284672 6.291017 6.303755 6.318636 6.34091 6.34151 6.343354 6.353502 6.368403 6.389339 6.390386 6.392929 6.413605 6.421451 6.44141 6.455884 6.476264 6.50008 6.511319 6.567543 6.569374 6.599825 6.611091 6.626693 6.631571 6.641151 6.662681 6.674827 6.677035 6.680493 6.700589 6.70191 6.704802 6.713351 6.713386 6.743378 6.755373 6.762756 6.787863 6.79495 6.815227 6.822907 6.828748 6.847308 6.851091 6.866085 6.895448 6.902334 6.933922 6.965888 6.969696 7.009603 7.022347 7.07351 7.08788 7.102858 7.112988 7.148097 7.149019 7.154397 7.161344 7.173941 7.174592 7.190528 7.21341 7.22951 7.233125 7.240516 7.246292 7.260381 7.261768 7.27143 7.27181 7.279017 7.280236 7.284534 7.285836 7.2958 7.314256 7.359199 7.361699 7.365365 7.365976 7.387201 7.408546 7.426832 7.458151 7.463198 7.483526 7.509777 7.598934 7.602959 7.608129 7.628581 7.633617 7.65376 7.654965 7.670688 7.683411 7.709543 7.735524 7.735765 7.739308 7.761287 7.793911 7.796098 7.800674 7.819169 7.833425 7.850186 7.86732 7.872321 7.874418 7.906221 7.927421 7.965047 7.965923 8.015361 8.023017 8.02532 8.026312 8.039224 8.042423 8.049236 8.075852 8.080662 8.082208 8.091076 8.12468 8.1282 8.128275 8.131603 8.133671 8.136946 8.145599 8.167336 8.181407 8.186605 8.187058 8.196745 8.208969 8.211716 8.211726 8.216777 8.22087 8.225588 8.290232 8.320789 8.347639 8.356984 8.357169 8.357901 8.361469 8.366411 8.368644 8.370338 8.375276 8.388346 8.390965 8.392914 8.407228 8.427279 8.428692 8.429038 8.436845 8.440774 8.487182 8.512758 8.524788 8.533193 8.534417 8.569938 8.57598 8.60075 8.603834 8.605624 8.620393 8.622101 8.623876 8.637653 8.66463 8.673926 8.705977 8.706846 8.706846 8.723983 8.757361 8.784218 8.784549 8.822476 8.85516 8.859759 8.874924 8.889799 8.891194 8.902277 8.905562 8.905892 8.91273 8.933155 8.984221 9.004698 9.006097 9.023025 9.027071 9.032563 9.05475 9.088004 9.10567 9.115306 9.126963 9.128359 9.149379 9.163459 9.197789 9.202183 9.218099 9.225257 9.252628 9.260797 9.272459 9.275051 9.311959 9.31321 9.336238 9.342951 9.359599 9.365391 9.385121 9.388263 9.401052 9.402691 9.409301 9.44797 9.454258 9.471273 9.484237 9.487332 9.494535 9.511947 9.512251 9.522128 9.532957 9.569663 9.580176 9.598546 9.601393 9.603345 9.60409 9.604945 9.621195 9.636513 9.636857 9.647658 9.648163 9.649841 9.650581 9.681837 9.699952 9.713029 9.714479 9.729196 9.733168 9.735046 9.746451 9.812325 9.822254 9.823939 9.834282 9.846933 9.880932 9.881922 9.893391 9.921301 9.943767 9.949885 9.952451 9.967013 9.978821 10.02549 10.02797 10.04632 10.05322 10.05744 10.08882 10.14668 10.165 10.17294 10.1948 10.20318 10.20657 10.20911 10.22917 10.23904 10.24966 10.25514 10.27454 10.3147 10.33212 10.34358 10.36856 10.36933 10.37567 10.38125 10.4171 10.43392 10.43443 10.49207 10.49473 10.49661 10.50908 10.51035 10.51762 10.58092 10.58287 10.6032 10.63406 10.64539 10.64781 10.65557 10.65891 10.68951 10.70175 10.70932 10.73333 10.74224 10.74419 10.77434 10.78518 10.79453 10.81194 10.82132 10.83333 10.85893 10.87079 10.95617 10.97329 10.99659 11.01384 11.01484 11.01881 11.03136 11.04846 11.06247 11.08848 11.10059 11.13929 11.15269 11.17807 11.20041 11.20131 11.20989 11.23355 11.24074 11.27037 11.27996 11.2924 11.29756 11.30326 11.31287 11.3164 11.34283 11.35485 11.3589 11.37105 11.37391 11.37598 11.39475 11.42868 11.43826 11.4404 11.44844 11.45421 11.45473 11.47992 11.48236 11.48242 11.49001 11.4901 11.52835 11.53496 11.53716 11.59932 11.60011 11.61487 11.61649 11.61758 11.63192 11.69098 11.69433 11.69822 11.72687 11.74551 
#> 

## Example 2: multinomial distribution:
mixt1 <- twoComp_mixt(n = 1500, weight = 0.8,
                      comp.dist = list("multinom", "multinom"),
                      comp.param = list(list("size" = 1, "prob" = c(0.2,0.3,0.5)),
                                        list("size" = 1, "prob" = c(0.1,0.6,0.3))))
data1 <- getmixtData(mixt1)
mixt2 <- twoComp_mixt(n = 2000, weight = 0.3,
                      comp.dist = list("multinom", "multinom"),
                      comp.param = list(list("size" = 1, "prob" = c(0.2,0.3,0.5)),
                                        list("size" = 1, "prob" = c(0.7,0.1,0.2))))
data2 <- getmixtData(mixt2)

## Define the admixture models:
admixMod1 <- admix_model(knownComp_dist = mixt1$comp.dist[[2]],
                         knownComp_param = mixt1$comp.param[[2]])
admixMod2 <- admix_model(knownComp_dist = mixt2$comp.dist[[2]],
                         knownComp_param = mixt2$comp.param[[2]])

## Estimate the mixture weights of the two admixture models (provide only hat(theta)_n):
estim_IBM(samples = list(data1,data2), admixMod = list(admixMod1,admixMod2))
#> Call:estim_IBM(samples = list(data1, data2), admixMod = list(admixMod1, 
#>     admixMod2))
#> 
#> Estimated mixing proportion (of the unknown component) in the 1st sample:  0.7824773 
#> Estimated mixing proportion (of the unknown component) in the 2nd sample:  0.3389325 
#> 
#> Fixed value for the mixing weight in the 1st sample (case of identical known components in the 2 samples): 
#> 
#> 
#> Support of integration: 1 2 3 
#>