Extractor for pairwise discrepancy rankings

Provide the matrix storing the ranks of discrepancies using Inversion-Best Matching approach between all couples among the K (K>2) samples under study.

Usage

get_discrepancy_rank(x)

Arguments

x

An object of class IBM_test.

Value

A matrix of ranks, from the closest couple (rank 1) in terms of discrepancy measure to the most different one.

Author

Xavier Milhaud xavier.milhaud.research@gmail.com

Examples

mixt1 <- twoComp_mixt(n = 380, weight = 0.7,
                      comp.dist = list("norm", "norm"),
                      comp.param = list(list("mean" = -2, "sd" = 0.5),
                                        list("mean" = 0, "sd" = 1)))
mixt2 <- twoComp_mixt(n = 350, weight = 0.85,
                      comp.dist = list("norm", "norm"),
                      comp.param = list(list("mean" = -2, "sd" = 0.5),
                                        list("mean" = -1, "sd" = 1)))
mixt3 <- twoComp_mixt(n = 500, weight = 0.6,
                      comp.dist = list("gamma", "gamma"),
                      comp.param = list(list("shape" = 16, "scale" = 1/4),
                                        list("shape" = 12, "scale" = 1/2)))
data1 <- get_mixture_data(mixt1)
data2 <- get_mixture_data(mixt2)
data3 <- get_mixture_data(mixt3)
admixMod1 <- admix_model(knownComp_dist = mixt1$comp.dist[[2]],
                         knownComp_param = mixt1$comp.param[[2]])
admixMod2 <- admix_model(knownComp_dist = mixt2$comp.dist[[2]],
                         knownComp_param = mixt2$comp.param[[2]])
admixMod3 <- admix_model(knownComp_dist = mixt3$comp.dist[[2]],
                         knownComp_param = mixt3$comp.param[[2]])
x <- admix_test(samples = list(data1,data2,data3),
                admixMod = list(admixMod1,admixMod2,admixMod3),
                conf_level = 0.95, test_method = "icv", n_sim_tab = 10)
get_discrepancy_id(x)
#>      [,1] [,2]  [,3] 
#> [1,] NA   "1-2" "1-3"
#> [2,] NA   NA    "2-3"
#> [3,] NA   NA    NA   
get_discrepancy_rank(x)
#>      [,1] [,2] [,3]
#> [1,]   NA    1    2
#> [2,]   NA   NA    3
#> [3,]   NA   NA   NA
get_discrepancy_matrix(x)
#>      [,1]       [,2]     [,3]
#> [1,]   NA 0.02578381 68.08923
#> [2,]   NA         NA 87.06412
#> [3,]   NA         NA       NA
get_statistic_components(x)
#> [1] "d_1-2 + d_1-3 + d_2-3"