Equality test of two unknown component distributions using polynomial expansions

Tests the null hypothesis (H0: f1=f2) using the decomposition of unknown component densities of two admixture distributions in an adequate orthonormal polynomial basis. Recall that we have two admixture models with respective probability density functions (pdf) l1 = p1*f1 + (1-p1)g1 and l2 = p2f2 + (1-p2)*g2, where g1 and g2 are the only known elements and l1 and l2 are observed. The admixture weights p1 and p2 thus have to be estimated. For further information on this method, see 'Details' below.

Usage

orthobasis_test(
  samples,
  admixMod,
  K = 3,
  s = 0.49,
  est_method = c("BVdk", "PS"),
  conf_level = 0.95,
  nb_echBoot = NULL,
  bounds_supp = NULL,
  support = c("Real", "Integer", "Positive", "Bounded.continuous", "Bounded.discrete")
)

Arguments

samples

(List) List of the two samples, each one following the mixture distribution given by l = p*f + (1-p)*g, with f and p unknown and g known.

admixMod

An object of class 'admix_model', containing useful information about distributions and parameters.

K

Number of coefficients considered for the polynomial basis expansion.

s

Rate at which the normalization factor is set in the penalization rule for model selection (in ]0,1/2[), see 'Details'.

est_method

Estimation method to get the component weights, either 'PS' (Patra and Sen estimation) or 'BVdk' (Bordes and Vendekerkhove estimation). Choosing 'PS' requires to specify the number of bootstrap samples.

conf_level

The confidence level, default to 95 percent. Equals 1-alpha, where alpha is the level of the test (type-I error).

nb_echBoot

(default to NULL) Number of bootstrap samples, useful when choosing 'PS' estimation method.

bounds_supp

(default to NULL) useful if support = 'bounded', a list of minimum and maximum bounds, specified as following: list( list(min.f1,min.g1,min.f2,min.g2) , list(max.f1,max.g1,max.f2,max.g2) )

support

support of the densities under consideration, useful to choose the polynomial orthonormal basis.

Value

A list with six elements containing: 1) the rejection decision; 2) the p-value of the test; 3) the test statistic; 4) the variance-covariance matrix of the test statistic; 5) selected rank for testing, and 6) estimates of the two component weights.

References

Milhaud X, Pommeret D, Salhi Y, Vandekerkhove P (2022). “Semiparametric two-sample admixture components comparison test: The symmetric case.” Journal of Statistical Planning and Inference, 216, 135-150. ISSN 0378-3758, doi:10.1016/j.jspi.2021.05.010 .

Author

Xavier Milhaud xavier.milhaud.research@gmail.com

Examples

# \donttest{
###### Using BVdk estimation (valid if symmetric unknown component densities).
#### Under the null hypothesis H0.
mixt1 <- twoComp_mixt(n = 1700, weight = 0.25,
                      comp.dist = list("norm", "norm"),
                      comp.param = list(list("mean" = 1, "sd" = 1),
                                        list("mean" = 4, "sd" = 1)))
data1 <- getmixtData(mixt1)
admixMod1 <- admix_model(knownComp_dist = mixt1$comp.dist[[2]],
                        knownComp_param = mixt1$comp.param[[2]])

mixt2 <- twoComp_mixt(n = 2000, weight = 0.15,
                      comp.dist = list("norm", "norm"),
                      comp.param = list(list("mean" = 1, "sd" = 1),
                                        list("mean" = 5, "sd" = 0.5)))
data2 <- getmixtData(mixt2)
admixMod2 <- admix_model(knownComp_dist = mixt2$comp.dist[[2]],
                        knownComp_param = mixt2$comp.param[[2]])

ortho_test <- orthobasis_test(samples = list(data1,data2),
                              admixMod = list(admixMod1,admixMod2),
                              K = 3, s = 0.49, nb_echBoot = NULL, support = 'Real',
                              bounds_supp = NULL, est_method = 'BVdk')
#>    Remind that choosing 'BVdk' for the estimation
#>     method is consistent only if the unknown component distribution
#>     has a symmetric density function.
# }