Print the results of estimated parameters from K admixture models

Print the estimated weight p of the unknown component in the admixture model under study Recall that an admixture model follows the cumulative distribution function (CDF) L, where L = p*F + (1-p)*G, with g a known CDF and p and f unknown quantities.

# S3 method for admix_estim
print(x, ...)

Arguments

x

An object of class 'admix_estim' (see ?admix_estim).

...

further arguments passed to or from other methods.

Author

Xavier Milhaud xavier.milhaud.research@gmail.com

Examples

##### On a simulated example to see whether the true parameters are well estimated.
list.comp <- list(f1 = "norm", g1 = "norm",
                  f2 = "norm", g2 = "norm")
list.param <- list(f1 = list(mean = 0, sd = 1), g1 = list(mean = 2, sd = 0.7),
                   f2 = list(mean = 0, sd = 1), g2 = list(mean = -3, sd = 1.1))
## Simulate data:
sim1 <- rsimmix(n = 2100, unknownComp_weight = 0.8, comp.dist = list(list.comp$f1,list.comp$g1),
                comp.param = list(list.param$f1, list.param$g1))$mixt.data
sim2 <- rsimmix(n= 2000, unknownComp_weight = 0.85, comp.dist = list(list.comp$f2,list.comp$g2),
                comp.param = list(list.param$f2, list.param$g2))$mixt.data
## Estimate the mixture weights of the admixture models:
list.comp <- list(f1 = NULL, g1 = "norm",
                  f2 = NULL, g2 = "norm")
list.param <- list(f1 = NULL, g1 = list(mean = 2, sd = 0.7),
                   f2 = NULL, g2 = list(mean = -3, sd = 1.1))
estim <- admix_estim(samples = list(sim1,sim2), sym.f = TRUE, est.method = 'IBM',
                     comp.dist = list.comp, comp.param = list.param)
#> Warning: Do not forget that estimators of proportions are reliable only if unknown component distributions are tested equal!
print(x = estim)
#> Call:
#> admix_estim(samples = list(sim1, sim2), sym.f = TRUE, est.method = "IBM", 
#>     comp.dist = list.comp, comp.param = list.param)
#> 
#> The number of populations/samples under study is 2.
#> 
#> 
#> The list of estimated weight(s) for the unknown component distribution(s) is :
#>    - estimated weight of the unknown component distribution for population  1 :  0.795919716768549
#>   - estimated weight of the unknown component distribution for population  2 :  0.84293075206806
#> 
#> The list of estimated location(s) for the unknown component distribution(s) is :
#>    - estimated location of the unknown component distribution for population  1 :  NA
#>   - estimated location of the unknown component distribution for population  2 :  NA
#> 
#> The chosen estimation technique is IBM.
#> Was the unknown density assumed to be symmetric (not important unless BVdk estimation is performed)? TRUE.
#>